Support event binding on shadow DOM elements
Previously elements using the shadow DOM broke when adding events.
This was because we use `document.evaluate` to find elements with
attributes starting with `@`, and the shadow DOM does not support
`document.evaluate`.

This commit adds support for shadow DOM elements to dig through all
children manually, which I expect to be slower but it is what it is.

I still need to benchmark it to confirm it's slower, if it's not then we
may as well do that for light DOM elements too.