Change template replacer syntax
We used to use `${var}` for the template replacement tags but this is
the same syntax used by JavaScript template literals, which could cause
some issues when generating the element body using JavaScript.

The new syntax is just `{var}`, this is compatible with JavaScript and
Jinja.

In the future I want this to be easily configurable by the user so we don't need to
worry about interop with other templating languages.