Add caching support to `DynamicFrameRouter`
Enabled by setting the `:caching` argument on the router, it will cache
the different frames and if a cached frame is loaded again it will be
fetched from the cache instead.

Changes to the DOM are persisted in the cache as this uses [cloneNode](https://developer.mozilla.org/en-US/docs/Web/API/Node/cloneNode).

The only problem I know of with this approach is that event listeners
are _not_ cloned.

We can maybe fix this ourselves by wrapping `addEventListener`,
apparently this is how jQuery handled it ([source](https://stackoverflow.com/a/15411683/15970183)).